+++ projects/packages/masterbar/src/admin-menu/class-atomic-admin-menu.php
@@ -328,14 +328,18 @@ class Atomic_Admin_Menu extends Admin_Menu {
 			parent::add_jetpack_menu();
 		}
 
+		$scan_position = $this->get_submenu_item_count( 'jetpack' ) - 1;
+
 		global $submenu;
-		$backup_submenu_label = __( 'Backup', 'jetpack-masterbar' );
-		$submenu_labels       = array_column( $submenu['jetpack'], 3 );
-		$backup_position      = array_search( $backup_submenu_label, $submenu_labels, true );
-		$scan_position        = $backup_position !== false ? $backup_position + 1 : $this->get_submenu_item_count( 'jetpack' ) - 1;
+		if ( isset( $submenu['jetpack'] ) ) {
+			$backup_submenu_label = __( 'Backup', 'jetpack-masterbar' );
+			$submenu_labels       = array_column( $submenu['jetpack'], 3 );
+			$backup_position      = array_search( $backup_submenu_label, $submenu_labels, true );
+			$scan_position        = $backup_position !== false ? $backup_position + 1 : $this->get_submenu_item_count( 'jetpack' ) - 1;
+		}
 
 		// @phan-suppress-next-line PhanTypeMismatchArgumentProbablyReal -- Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
-		add_submenu_page( 'jetpack', esc_attr__( 'Scan', 'jetpack-masterbar' ), __( 'Scan', 'jetpack-masterbar' ), 'manage_options', 'https://wordpress.com/scan/history/' . $this->domain, null, $scan_position );
+		add_submenu_page( 'jetpack', esc_attr__( 'Scan', 'jetpack-masterbar' ), __( 'Scan', 'jetpack-masterbar' ), 'manage_options', 'https://wordpress.com/scan/' . $this->domain, null, $scan_position );
 
 		/**
 		 * Prevent duplicate menu items that link to Jetpack Backup.