+++ projects/plugins/jetpack/_inc/client/newsletter/email-settings.jsx
@@ -18,9 +18,10 @@ import SettingsGroup from 'components/settings-group';
 import SupportInfo from 'components/support-info';
 import TextInput from 'components/text-input';
 import analytics from 'lib/analytics';
+import { FEATURE_NEWSLETTER_JETPACK } from 'lib/plans/constants';
 import { useCallback, useState } from 'react';
 import { connect } from 'react-redux';
-import { isUnavailableInOfflineMode, isUnavailableInSiteConnectionMode } from 'state/connection';
+import { isUnavailableInOfflineMode, hasConnectedOwner } from 'state/connection';
 import {
 	getSiteTitle,
 	getUserGravatar,
@@ -58,17 +59,17 @@ const EmailSettings = props => {
 		subscriptionReplyTo,
 		subscriptionFromName,
 		updateFormStateAndSaveOptionValue,
-		unavailableInSiteConnectionMode,
 		gravatar,
 		email,
 		adminUrl,
 		displayName,
 		dateExample,
 		siteName,
+		siteHasConnectedUser,
 	} = props;
 
-	const disabled =
-		! isSubscriptionsActive || unavailableInOfflineMode || unavailableInSiteConnectionMode;
+	const disabled = ! siteHasConnectedUser || ! isSubscriptionsActive || unavailableInOfflineMode;
+	const isSaving = isSavingAnyOption( [ GRAVATER_OPTION, AUTHOR_OPTION, POST_DATE_OPTION ] );
 	const gravatarInputDisabled = disabled || isSavingAnyOption( [ GRAVATER_OPTION ] );
 	const authorInputDisabled = disabled || isSavingAnyOption( [ AUTHOR_OPTION ] );
 	const postDateInputDisabled = disabled || isSavingAnyOption( [ POST_DATE_OPTION ] );
@@ -179,13 +180,15 @@ const EmailSettings = props => {
 			{ ...props }
 			header={ __( 'Email configuration', 'jetpack' ) }
 			hideButton
+			feature={ FEATURE_NEWSLETTER_JETPACK }
 			module={ SUBSCRIPTIONS_MODULE_NAME }
-			saveDisabled={ disabled }
+			saveDisabled={ isSaving }
+			isDisabled={ disabled }
 		>
 			<SettingsGroup
 				hasChild
 				disableInOfflineMode
-				disableInSiteConnectionMode
+				disableInSiteConnectionMode={ ! siteHasConnectedUser }
 				module={ subscriptionsModule }
 				support={ {
 					link: featuredImageInEmailSupportUrl,
@@ -210,7 +213,7 @@ const EmailSettings = props => {
 			<SettingsGroup
 				hasChild
 				disableInOfflineMode
-				disableInSiteConnectionMode
+				disableInSiteConnectionMode={ ! siteHasConnectedUser }
 				module={ subscriptionsModule }
 				className="newsletter-group"
 			>
@@ -319,7 +322,7 @@ const EmailSettings = props => {
 			<SettingsGroup
 				hasChild
 				disableInOfflineMode
-				disableInSiteConnectionMode
+				disableInSiteConnectionMode={ ! siteHasConnectedUser }
 				module={ subscriptionsModule }
 				support={ {
 					link: subscriptionsAndNewslettersSupportUrl,
@@ -357,7 +360,7 @@ const EmailSettings = props => {
 			<SettingsGroup
 				hasChild
 				disableInOfflineMode
-				disableInSiteConnectionMode
+				disableInSiteConnectionMode={ ! siteHasConnectedUser }
 				module={ subscriptionsModule }
 				className="newsletter-group"
 				support={ {
@@ -407,17 +410,9 @@ const EmailSettings = props => {
 				</p>
 				<RadioControl
 					className="jp-form-radio-gap"
-					selected={ subscriptionReplyTo || 'no-reply' }
+					selected={ subscriptionReplyTo || 'comment' }
 					disabled={ replyToInputDisabled }
 					options={ [
-						{
-							label: (
-								<span className="jp-form-toggle-explanation">
-									{ __( 'Replies are not allowed', 'jetpack' ) }
-								</span>
-							),
-							value: 'no-reply',
-						},
 						{
 							label: (
 								<span className="jp-form-toggle-explanation">
@@ -433,6 +428,14 @@ const EmailSettings = props => {
 								</span>
 							),
 							value: 'author',
+						},
+						{
+							label: (
+								<span className="jp-form-toggle-explanation">
+									{ __( 'Replies are not allowed', 'jetpack' ) }
+								</span>
+							),
+							value: 'no-reply',
 						},
 					] }
 					onChange={ handleSubscriptionReplyToChange }
@@ -475,10 +478,7 @@ export default withModuleSettingsFormHelpers(
 			subscriptionFromName: ownProps.getOptionValue( FROM_NAME_OPTION ),
 			dateExample: getNewsetterDateExample( state ),
 			unavailableInOfflineMode: isUnavailableInOfflineMode( state, SUBSCRIPTIONS_MODULE_NAME ),
-			unavailableInSiteConnectionMode: isUnavailableInSiteConnectionMode(
-				state,
-				SUBSCRIPTIONS_MODULE_NAME
-			),
+			siteHasConnectedUser: hasConnectedOwner( state ),
 		};
 	} )( EmailSettings )
 );