+++ projects/plugins/jetpack/_inc/client/traffic/blaze.jsx
@@ -1,7 +1,6 @@
 import { getRedirectUrl, ToggleControl } from '@automattic/jetpack-components';
 import { __ } from '@wordpress/i18n';
 import Card from 'components/card';
-import ConnectUserBar from 'components/connect-user-bar';
 import { withModuleSettingsFormHelpers } from 'components/module-settings/with-module-settings-form-helpers';
 import { ModuleToggle } from 'components/module-toggle';
 import SettingsCard from 'components/settings-card';
@@ -15,6 +14,7 @@ import {
 	shouldInitializeBlaze,
 } from 'state/initial-state';
 import { getModule } from 'state/modules';
+import { FEATURE_JETPACK_BLAZE } from '../lib/plans/constants';
 
 const trackDashboardClick = () => {
 	analytics.tracks.recordJetpackClick( 'blaze-dashboard' );
@@ -41,6 +41,8 @@ function Blaze( props ) {
 		toggleModuleNow,
 	} = props;
 
+	const { can_init: canInit, reason } = blazeAvailable;
+
 	if ( isWoASite && ! blazeDashboardEnabled ) {
 		return null;
 	}
@@ -66,7 +68,16 @@ function Blaze( props ) {
 	};
 
 	const blazeToggle = () => {
-		if ( ! blazeAvailable ) {
+		if ( ! canInit && reason === 'user_not_connected' ) {
+			return (
+				<ToggleControl
+					disabled={ true }
+					label={ __( 'Attract high-quality traffic to your site using Blaze.', 'jetpack' ) }
+				/>
+			);
+		}
+
+		if ( ! canInit ) {
 			return (
 				<ToggleControl
 					disabled={ true }
@@ -91,7 +102,13 @@ function Blaze( props ) {
 	};
 
 	return (
-		<SettingsCard { ...props } header={ __( 'Blaze', 'jetpack' ) } module="blaze" hideButton>
+		<SettingsCard
+			{ ...props }
+			header={ __( 'Blaze', 'jetpack' ) }
+			module="blaze"
+			hideButton
+			feature={ FEATURE_JETPACK_BLAZE }
+		>
 			<SettingsGroup
 				module={ { module: 'blaze' } }
 				disableInOfflineMode
@@ -103,18 +120,7 @@ function Blaze( props ) {
 			>
 				{ blazeToggle() }
 			</SettingsGroup>
-			{ blazeAvailable &&
-				blazeActive &&
-				hasConnectedOwner &&
-				! isOfflineMode &&
-				blazeDashboardLink() }
-			{ blazeAvailable && ! hasConnectedOwner && ! isOfflineMode && (
-				<ConnectUserBar
-					feature="blaze"
-					featureLabel={ __( 'Blaze', 'jetpack' ) }
-					text={ __( 'Connect to set up campaigns and promote your content.', 'jetpack' ) }
-				/>
-			) }
+			{ canInit && blazeActive && ! isOfflineMode && blazeDashboardLink() }
 		</SettingsCard>
 	);
 }