+++ projects/plugins/jetpack/tests/php/general/test_class.jetpack.php
@@ -200,64 +200,6 @@ EXPECTED;
 		$this->assertEquals( $expected, $result );
 	}
 
-	/**
-	 * @author tonykova
-	 */
-	public function test_implode_frontend_css_enqueues_bundle_file_handle() {
-		global $wp_styles;
-		$wp_styles = new WP_Styles();
-
-		add_filter( 'jetpack_implode_frontend_css', '__return_true' );
-
-		if ( ! file_exists( plugins_url( 'jetpack-carousel.css', __FILE__ ) ) ) {
-			$this->markTestSkipped( 'Required CSS file not found.' );
-		}
-
-		// Enqueue some script on the $to_dequeue list
-		$style_handle = 'jetpack-carousel';
-		wp_enqueue_style( 'jetpack-carousel', plugins_url( 'jetpack-carousel.css', __FILE__ ) ); // phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
-
-		Jetpack::init()->implode_frontend_css( true );
-
-		$seen_bundle = false;
-		foreach ( $wp_styles->registered as $handle => $handle_obj ) {
-			if ( $style_handle === $handle ) {
-				$expected = ( defined( 'WP_DEBUG' ) && WP_DEBUG ) ? "<!-- `{$style_handle}` is included in the concatenated jetpack.css -->\r\n" : '';
-				$this->assertEquals( $expected, get_echo( array( $wp_styles, 'do_item' ), array( $handle ) ) );
-			} elseif ( 'jetpack_css' === $handle ) {
-				$seen_bundle = true;
-			}
-		}
-
-		$this->assertTrue( $seen_bundle );
-	}
-
-	/**
-	 * @author tonykova
-	 * @since 3.2.0
-	 */
-	public function test_implode_frontend_css_does_not_enqueue_bundle_when_disabled_through_filter() {
-		global $wp_styles;
-		$wp_styles = new WP_Styles();
-
-		add_filter( 'jetpack_implode_frontend_css', '__return_false' );
-
-		// Enqueue some script on the $to_dequeue list
-		wp_enqueue_style( 'jetpack-carousel', plugins_url( 'jetpack-carousel.css', __FILE__ ) ); // phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
-
-		Jetpack::init()->implode_frontend_css();
-
-		$seen_orig = false;
-		foreach ( $wp_styles->registered as $handle => $handle_obj ) {
-			$this->assertNotEquals( 'jetpack_css', $handle );
-			if ( 'jetpack-carousel' === $handle ) {
-				$seen_orig = true;
-			}
-		}
-
-		$this->assertTrue( $seen_orig );
-	}
-
 	public function test_activating_deactivating_modules_fires_actions() {
 		self::reset_tracking_of_module_activation();