+++ projects/js-packages/base-styles/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.33] - 2024-10-07
+### Changed
+- Updated package dependencies. [#39594]
+
 ## [0.6.32] - 2024-09-10
 ### Changed
 - Updated package dependencies. [#39302]
@@ -313,6 +317,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies.
 - Use Node 16.7.0 in tooling. This shouldn't change the behavior of the code itself.
 
+[0.6.33]: https://github.com/Automattic/jetpack-base-styles/compare/0.6.32...0.6.33
 [0.6.32]: https://github.com/Automattic/jetpack-base-styles/compare/0.6.31...0.6.32
 [0.6.31]: https://github.com/Automattic/jetpack-base-styles/compare/0.6.30...0.6.31
 [0.6.30]: https://github.com/Automattic/jetpack-base-styles/compare/0.6.29...0.6.30