+++ projects/packages/masterbar/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.5] - 2024-10-07
+### Changed
+- Updated package dependencies. [#39594]
+- Update Jetpack Scan link. [#39619]
+
+## [0.9.4] - 2024-09-30
+### Changed
+- Remove user connection nudges where they aren't needed. Add user connection nudges where needed [#39533]
+
+### Removed
+- Masterbar: Remove User Info side-panel [#39546]
+
+### Fixed
+- Admin bar: don't enqueue obsolete Core CSS overrides for Default scheme [#39453]
+
 ## [0.9.3] - 2024-09-23
 ### Changed
 - Update dependencies.
@@ -130,6 +145,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies. [#37669]
 - Updated package dependencies. [#37706]
 
+[0.9.5]: https://github.com/Automattic/jetpack-masterbar/compare/v0.9.4...v0.9.5
+[0.9.4]: https://github.com/Automattic/jetpack-masterbar/compare/v0.9.3...v0.9.4
 [0.9.3]: https://github.com/Automattic/jetpack-masterbar/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/Automattic/jetpack-masterbar/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/Automattic/jetpack-masterbar/compare/v0.9.0...v0.9.1