+++ projects/packages/wordads/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.35] - 2024-10-07
+### Changed
+- Updated package dependencies. [#39592]
+- Updated package dependencies. [#39594]
+- Updated package dependencies. [#39639]
+
+## [0.3.34] - 2024-09-30
+### Changed
+- Update dependencies. [#39528]
+
 ## [0.3.33] - 2024-09-23
 ### Changed
 - Update dependencies. [#38017]
@@ -412,6 +422,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PHPCS: Fix `WordPress.Security.ValidatedSanitizedInput`
 - Updated package dependencies.
 
+[0.3.35]: https://github.com/Automattic/jetpack-wordads/compare/v0.3.34...v0.3.35
+[0.3.34]: https://github.com/Automattic/jetpack-wordads/compare/v0.3.33...v0.3.34
 [0.3.33]: https://github.com/Automattic/jetpack-wordads/compare/v0.3.32...v0.3.33
 [0.3.32]: https://github.com/Automattic/jetpack-wordads/compare/v0.3.31...v0.3.32
 [0.3.31]: https://github.com/Automattic/jetpack-wordads/compare/v0.3.30...v0.3.31